+++ lib/contentful/management/resource/refresher.rb
@@ -6,9 +6,13 @@ module Contentful
         # Reload an object
         # Updates the current version of the object to the version on the system
         def reload
-          self_class = self.class
-          resource = self.is_a?(Space) ? self_class.find(client, id) : self_class.find(client, space.id, id)
-          refresh_data(resource) if resource.is_a? self_class
+          resource = refresh_find
+          refresh_data(resource) if resource.is_a? self.class
+        end
+
+        # @private
+        def refresh_find
+          self.class.find(client, space.id, id)
         end
 
         # @private